+++ app/src/main/java/org/simple/clinic/widgets/qrcodescanner/MLKitQrCodeAnalyzer.kt
@@ -3,6 +3,8 @@ package org.simple.clinic.widgets.qrcodescanner
 import android.annotation.SuppressLint
 import androidx.camera.core.ImageAnalysis
 import androidx.camera.core.ImageProxy
+import com.google.android.gms.dynamite.DynamiteModule.LoadingException
+import com.google.mlkit.common.MlKitException
 import com.google.mlkit.vision.barcode.Barcode
 import com.google.mlkit.vision.barcode.BarcodeScannerOptions
 import com.google.mlkit.vision.barcode.BarcodeScanning
@@ -10,14 +12,14 @@ import com.google.mlkit.vision.common.InputImage
 
 class MLKitQrCodeAnalyzer(
     private val bitmapUtils: BitmapUtils,
-    private val onQrCodeDetected: OnQrCodeDetected
+    private val onQrCodeDetected: OnQrCodeDetected,
+    private val mlKitUnavailable: () -> Unit
 ) : ImageAnalysis.Analyzer {
 
   private val options = BarcodeScannerOptions.Builder()
       .setBarcodeFormats(Barcode.FORMAT_QR_CODE)
       .build()
   private val scanner = BarcodeScanning.getClient(options)
-
 
   @SuppressLint("UnsafeExperimentalUsageError")
   override fun analyze(imageProxy: ImageProxy) {
@@ -41,11 +43,16 @@ class MLKitQrCodeAnalyzer(
             }
           }
         }
-        .addOnFailureListener {
-          // Do nothing
-        }
+        .addOnFailureListener(::handleException)
         .addOnCompleteListener {
           imageProxy.close()
         }
+  }
+
+  private fun handleException(exception: Exception) {
+    val isMLKitUnavailable = exception is MlKitException && exception.errorCode == MlKitException.UNAVAILABLE
+    if (exception is LoadingException || isMLKitUnavailable) {
+      mlKitUnavailable()
+    }
   }
 }